+++ react/QueryProducts.tsx
@@ -1,4 +1,5 @@
-import products from './queries/products.gql'
 import { DocumentNode } from 'graphql'
+
+import products from './queries/products.gql'
 
 export default products as DocumentNode